reorganize the folder structure: move lialg and kernelregression into internal, move rest to pkg
