+++ internal/lialg/eigen.go
