+++ internal/lialg/eigen_test.go
