+++ internal/lialg/qpsolver.go
