+++ internal/lialg/qpsolver_test.go
