+++ internal/lialg/rbfkernel.go
