+++ internal/lialg/rbfkernel_test.go
